requirements.txt
absl-py>=0.7.0
gym==0.17.3
numpy>=1.18.5
pybullet>=3.0.4
matplotlib>=3.1.1
opencv-python>=4.1.2.30
meshcat>=0.0.18
scipy==1.4.1
scikit-image>=0.17.2
scikit-learn>=1.0.2
transforms3d==0.3.1
tensorflow==2.8.0
tensorflow-hub==0.12.0